Revert "Break down ODistributedLifecycleListener"
