+++ distributed-v2/src/test/java/com/orientechnologies/orient/server/distributed/DistributedLifecycleListenerIT.java
@@ -22,8 +22,6 @@ package com.orientechnologies.orient.server.distributed;
 
 import com.orientechnologies.common.log.OLogManager;
 import com.orientechnologies.common.util.OPair;
-import com.orientechnologies.orient.server.distributed.listener.ODistributedDatabaseStatusChangeListener;
-import com.orientechnologies.orient.server.distributed.listener.ODistributedNodeLifecycleListener;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -35,9 +33,7 @@ import java.util.concurrent.atomic.AtomicLong;
 /**
  * Tests the behavior of hooks in distributed configuration.
  */
-public class DistributedLifecycleListenerIT extends AbstractServerClusterTest
-    implements ODistributedNodeLifecycleListener, ODistributedDatabaseStatusChangeListener {
-
+public class DistributedLifecycleListenerIT extends AbstractServerClusterTest implements ODistributedLifecycleListener {
   private final static int                                               SERVERS        = 2;
 
   private final AtomicLong                                               beforeNodeJoin = new AtomicLong();
@@ -83,9 +79,7 @@ public class DistributedLifecycleListenerIT extends AbstractServerClusterTest
 
   @Override
   protected void onServerStarted(ServerRun server) {
-    server.getServerInstance().getDistributedManager()
-        .registerDistributedNodeLifecycleListener(this)
-        .registerDistributedDatabaseStatusChangeListener(this);
+    server.getServerInstance().getDistributedManager().registerLifecycleListener(this);
   }
 
   @Override
