+++ distributed-v2/src/test/java/com/orientechnologies/orient/server/distributed/HaRemoveServerIT.java
@@ -2,7 +2,6 @@ package com.orientechnologies.orient.server.distributed;
 
 import com.orientechnologies.orient.core.db.document.ODatabaseDocument;
 import com.orientechnologies.orient.core.sql.OCommandSQL;
-import com.orientechnologies.orient.server.distributed.listener.ODistributedDatabaseStatusChangeListener;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -63,7 +62,22 @@ public class HaRemoveServerIT extends AbstractServerClusterTest {
 
       // 3 events for failing test (NOT_AVAILABLE,SYNCHRONIZING,ONLINE)
       final CountDownLatch latch = new CountDownLatch(3);
-      distributedManager.registerDistributedDatabaseStatusChangeListener(new ODistributedDatabaseStatusChangeListener() {
+      distributedManager.registerLifecycleListener(new ODistributedLifecycleListener() {
+        @Override
+        public boolean onNodeJoining(String iNode) {
+          return false;
+        }
+
+        @Override
+        public void onNodeJoined(String iNode) {
+
+        }
+
+        @Override
+        public void onNodeLeft(String iNode) {
+
+        }
+
         @Override
         public void onDatabaseChangeStatus(String iNode, String iDatabaseName, ODistributedServerManager.DB_STATUS iNewStatus) {
