+++ distributed-v2/src/test/java/com/orientechnologies/orient/server/distributed/StopNodeIT.java
@@ -18,7 +18,6 @@ package com.orientechnologies.orient.server.distributed;
 import com.orientechnologies.common.log.OLogManager;
 import com.orientechnologies.common.util.OCallable;
 import com.orientechnologies.orient.core.db.document.ODatabaseDocument;
-import com.orientechnologies.orient.server.distributed.listener.ODistributedNodeLifecycleListener;
 import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -68,7 +67,7 @@ public class StopNodeIT extends AbstractServerClusterTxTest {
 
     if (serverStarted == 0) {
       // INSTALL ON FIRST SERVER ONLY THE SERVER MONITOR TO CHECK IF HAS BEEN RESTARTED
-      server.server.getDistributedManager().registerDistributedNodeLifecycleListener(new ODistributedNodeLifecycleListener() {
+      server.server.getDistributedManager().registerLifecycleListener(new ODistributedLifecycleListener() {
         @Override
         public boolean onNodeJoining(String iNode) {
           return true;
@@ -84,6 +83,10 @@ public class StopNodeIT extends AbstractServerClusterTxTest {
           nodeReJoined.clear();
           nodeLefts.incrementAndGet();
           OLogManager.instance().info(this, "NODE LEFT %s = %d", iNode, nodeLefts.get());
+        }
+
+        @Override
+        public void onDatabaseChangeStatus(String iNode, String iDatabaseName, ODistributedServerManager.DB_STATUS iNewStatus) {
         }
       });
     }
