+++ distributed/src/test/java/com/orientechnologies/orient/server/distributed/OneNodeBackupIT.java
@@ -16,7 +16,6 @@
 package com.orientechnologies.orient.server.distributed;
 
 import com.orientechnologies.orient.core.db.document.ODatabaseDocument;
-import com.orientechnologies.orient.server.distributed.listener.ODistributedNodeLifecycleListener;
 import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
@@ -63,7 +62,7 @@ public class OneNodeBackupIT extends AbstractServerClusterTxTest {
 
     if (serverStarted == 0) {
       // INSTALL ON FIRST SERVER ONLY THE SERVER MONITOR TO CHECK IF HAS BEEN RESTARTED
-      server.server.getDistributedManager().registerDistributedNodeLifecycleListener(new ODistributedNodeLifecycleListener() {
+      server.server.getDistributedManager().registerLifecycleListener(new ODistributedLifecycleListener() {
         @Override
         public boolean onNodeJoining(String iNode) {
           return true;
@@ -76,6 +75,10 @@ public class OneNodeBackupIT extends AbstractServerClusterTxTest {
         @Override
         public void onNodeLeft(String iNode) {
           nodeLefts.incrementAndGet();
+        }
+
+        @Override
+        public void onDatabaseChangeStatus(String iNode, String iDatabaseName, ODistributedServerManager.DB_STATUS iNewStatus) {
         }
       });
     }
