+++ distributed/src/test/resources/hazelcast-0.xml
@@ -9,7 +9,7 @@
 	the specific language governing permissions and ~ limitations under the License. -->
 
 <hazelcast
-        xsi:schemaLocation="http://www.hazelcast.com/schema/config hazelcast-config-3.10.xsd"
+        xsi:schemaLocation="http://www.hazelcast.com/schema/config hazelcast-config-3.3.xsd"
         xmlns="http://www.hazelcast.com/schema/config" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
     <group>
         <name>orientdb</name>
