--- server/src/main/java/com/orientechnologies/orient/server/distributed/listener/ODistributedDatabaseStatusChangeListener.java
@@ -1,33 +1,0 @@
-/*
- *
- *  *  Copyright 2010-2016 OrientDB LTD (http://orientdb.com)
- *  *
- *  *  Licensed under the Apache License, Version 2.0 (the "License");
- *  *  you may not use this file except in compliance with the License.
- *  *  You may obtain a copy of the License at
- *  *
- *  *       http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  *  Unless required by applicable law or agreed to in writing, software
- *  *  distributed under the License is distributed on an "AS IS" BASIS,
- *  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  *  See the License for the specific language governing permissions and
- *  *  limitations under the License.
- *  *
- *  * For more information: http://orientdb.com
- *
- */
-package com.orientechnologies.orient.server.distributed.listener;
-
-import com.orientechnologies.orient.server.distributed.ODistributedServerManager;
-
-public interface ODistributedDatabaseStatusChangeListener {
-  /**
-   * Called upon change of database status on a node. Available statuses are defined in ODistributedServerManager.DB_STATUS.
-   *
-   * @param iNode         The node name
-   * @param iDatabaseName Database name
-   * @param iNewStatus    The new status
-   */
-  void onDatabaseChangeStatus(String iNode, String iDatabaseName, ODistributedServerManager.DB_STATUS iNewStatus);
-}
